Split up Profile into multiple modules
